requirements.txt
appdirs==1.4.4
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
distlib==0.3.1
Django==3.1.7
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.2
filelock==3.0.12
idna==2.10
Markdown==3.3.4
pbr==5.5.1
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
stevedore==3.3.0
Unipath==1.1
urllib3==1.26.3
virtualenv==20.4.2
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
